ruh-backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0
groq==0.31.1
tenacity==8.2.3
numpy==1.24.3
sentence-transformers==2.2.2
python-dateutil==2.8.2